remove SonarCloud Code Smells
